Move file generation to new package
