+++ cmd/gomu/cmd/cli/new/new.go
@@ -4,12 +4,11 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"path/filepath"
 	"strings"
-	"text/template"
 
 	"github.com/asim/go-micro/cmd/gomu/cmd"
-	tmpl "github.com/asim/go-micro/cmd/gomu/cmd/cli/new/template"
+	"github.com/asim/go-micro/cmd/gomu/generate"
+	tmpl "github.com/asim/go-micro/cmd/gomu/generate/template"
 	"github.com/urfave/cli/v2"
 )
 
@@ -22,21 +21,6 @@ var flags []cli.Flag = []cli.Flag{
 		Name:  "skaffold",
 		Usage: "generate skaffold files",
 	},
-}
-
-type config struct {
-	Alias    string
-	Comments []string
-	Dir      string
-	Vendor   string
-	Client   bool
-	Jaeger   bool
-	Skaffold bool
-}
-
-type file struct {
-	Path string
-	Tmpl string
 }
 
 // NewCommand returns a new new cli command.
@@ -105,7 +89,7 @@ func createProject(ctx *cli.Context, pt string) error {
 
 	fmt.Printf("creating %s %s\n", pt, name)
 
-	files := []file{
+	files := []generate.File{
 		{".dockerignore", tmpl.DockerIgnore},
 		{".gitignore", tmpl.GitIgnore},
 		{"Dockerfile", tmpl.Dockerfile},
@@ -115,17 +99,17 @@ func createProject(ctx *cli.Context, pt string) error {
 
 	switch pt {
 	case "client":
-		files = append(files, []file{
+		files = append(files, []generate.File{
 			{"main.go", tmpl.MainCLT},
 		}...)
 	case "function":
-		files = append(files, []file{
+		files = append(files, []generate.File{
 			{"handler/" + name + ".go", tmpl.HandlerFNC},
 			{"main.go", tmpl.MainFNC},
 			{"proto/" + name + ".proto", tmpl.ProtoFNC},
 		}...)
 	case "service":
-		files = append(files, []file{
+		files = append(files, []generate.File{
 			{"handler/" + name + ".go", tmpl.HandlerSRV},
 			{"main.go", tmpl.MainSRV},
 			{"proto/" + name + ".proto", tmpl.ProtoSRV},
@@ -135,7 +119,7 @@ func createProject(ctx *cli.Context, pt string) error {
 	}
 
 	if ctx.Bool("skaffold") {
-		files = append(files, []file{
+		files = append(files, []generate.File{
 			{"plugins.go", tmpl.Plugins},
 			{"resources/clusterrole.yaml", tmpl.KubernetesClusterRole},
 			{"resources/configmap.yaml", tmpl.KubernetesEnv},
@@ -145,7 +129,7 @@ func createProject(ctx *cli.Context, pt string) error {
 		}...)
 	}
 
-	c := config{
+	c := generate.Config{
 		Alias:    name,
 		Dir:      dir,
 		Vendor:   vendor,
@@ -184,41 +168,8 @@ func protoComments(name, dir string) []string {
 	}
 }
 
-func create(files []file, c config) error {
-	for _, file := range files {
-		fp := filepath.Join(c.Dir, file.Path)
-		dir := filepath.Dir(fp)
-
-		if _, err := os.Stat(dir); os.IsNotExist(err) {
-			if err := os.MkdirAll(dir, 0755); err != nil {
-				return err
-			}
-		}
-
-		f, err := os.Create(fp)
-		if err != nil {
-			return err
-		}
-
-		fn := template.FuncMap{
-			"dehyphen": func(s string) string {
-				return strings.ReplaceAll(s, "-", "")
-			},
-			"lower": strings.ToLower,
-			"title": func(s string) string {
-				return strings.ReplaceAll(strings.Title(s), "-", "")
-			},
-		}
-		t, err := template.New(fp).Funcs(fn).Parse(file.Tmpl)
-		if err != nil {
-			return err
-		}
-
-		err = t.Execute(f, c)
-		if err != nil {
-			return err
-		}
-	}
+func create(files []generate.File, c generate.Config) error {
+	generate.Create(files, c)
 
 	for _, comment := range c.Comments {
 		fmt.Println(comment)
