+++ cmd/gomu/generate/config.go
@@ -1,0 +1,11 @@
+package generate
+
+type Config struct {
+	Alias    string
+	Comments []string
+	Dir      string
+	Vendor   string
+	Client   bool
+	Jaeger   bool
+	Skaffold bool
+}
