+++ cmd/gomu/generate/template/gitignore.go
