+++ cmd/gomu/generate/template/kubernetes.go
