+++ cmd/gomu/generate/template/main.go
