+++ cmd/gomu/generate/template/makefile.go
