+++ cmd/gomu/generate/template/module.go
