+++ cmd/gomu/generate/template/plugins.go
