+++ cmd/gomu/generate/template/proto.go
