+++ cmd/gomu/generate/template/skaffold.go
